Add files via upload
Front body part select screen. Have the framework, uses SVG to have image of body that scales to screen and has body part regions that act as clickable regions to navigate to next screen. The SVG is temporary and will be replaced with something more polished. Also, not all regions are implemented (like the wrist and the entire back side), and the region for the neck is kinda messed up but it's an easy fix, and I'll do that when I make the actual body SVG.