Finished front and back select page, all linked
Replaced body with a better defined image, separate "back" image, added front and back buttons and additional body groups.

Note: There are wrist groups drawn, but they link to the hands since there is no wrist page. Also, for the arms, B3=T3 and vice versa for some reason. So if we ever bother to fix that, the links in the svg should be changed as well.